+++ include/fluxgl/fluxgl.h
@@ -5,4 +5,5 @@
 #include "core/log.h"
 
 #include "graphics/vertex.h"
+#include "graphics/shader.h"
 #include "graphics/mesh.h"